Combine into one model file with causal lm outputs for forward
