+++ src/transformers/models/granite_speech/modeling_granite_speech.py
@@ -1,24 +1,329 @@
-from typing import List, Optional
+import math
+from dataclasses import dataclass
+from typing import List, Optional, Union, Tuple
 
 import torch
 import torch.utils.checkpoint
+from torch import nn, einsum
+import torch.nn.functional as F
 
+from ...modeling_outputs import BaseModelOutput, ModelOutput
+
+from transformers import Blip2QFormerModel
 from transformers.generation import GenerationMixin
 from transformers.modeling_utils import PreTrainedModel
 from transformers.models.granite import GraniteForCausalLM
-from .configuration_granite_speech import GraniteSpeechConfig
-from .projector import EncoderProjectorQFormer
-from .encoder import CTCModel
+from .configuration_granite_speech import (
+    GraniteSpeechConfig,
+    GraniteSpeechEncoderConfig,
+)
 
 from peft import get_peft_model, LoraConfig, TaskType
 import time
 
 
+@dataclass
+class GraniteSpeechCausalLMOutputWithPast(ModelOutput):
+    loss: Optional[torch.FloatTensor] = None
+    logits: torch.FloatTensor = None
+    past_key_values: Optional[List[torch.FloatTensor]] = None
+    hidden_states: Optional[Tuple[torch.FloatTensor]] = None
+    attentions: Optional[Tuple[torch.FloatTensor]] = None
+    attention_mask: Optional[torch.FloatTensor] = None
+
+
+### Projector
+class EncoderProjectorQFormer(nn.Module):
+    def __init__(self, config: GraniteSpeechConfig):
+        super().__init__()
+        self.hidden_size = config.hidden_size
+        self.ds_rate = config.downsample_rate
+        self.window_size = config.window_size
+        self.num_queries = self.window_size // self.ds_rate
+        self.query = nn.Parameter(torch.zeros(1, self.num_queries, config.hidden_size))
+        self.query.data.normal_(mean=0.0, std=1.0)
+        self.qformer = Blip2QFormerModel(config)
+        self.linear = nn.Linear(config.hidden_size, config.llm_dim)
+
+    def forward(self, x, atts):
+        batch_size, seq_len, dim = x.size()
+        nblocks = math.ceil(seq_len / self.window_size)
+        pad = nblocks * self.window_size - seq_len
+        x = nn.functional.pad(x, (0, 0, 0, pad), "constant", 0)
+        x = x.view(batch_size * nblocks, self.window_size, dim)
+
+        query_output = self.qformer(
+            query_embeds=self.query.data,
+            encoder_hidden_states=x,
+            encoder_attention_mask=atts,
+            return_dict=True,
+        )
+        query_proj = self.linear(
+            query_output.last_hidden_state.view(
+                batch_size, nblocks * self.window_size // self.ds_rate, -1
+            )
+        )
+        return query_proj
+
+### Encoder
+class CTCModel(nn.Module):
+    def __init__(self, config: GraniteSpeechEncoderConfig):
+        super(CTCModel, self).__init__()
+
+        self.rnn_trL = [nn.Linear(config.input_dim, config.hidden_dim, bias=True)]
+        for l in range(config.num_layers):
+            self.rnn_trL.append(
+                ConformerBlock(
+                    dim=config.hidden_dim,
+                    dim_head=config.dim_head,
+                    heads=config.num_heads,
+                    ff_mult=config.feedforward_mult,
+                    conv_expansion_factor=config.conv_expansion_factor,
+                    conv_kernel_size=config.conv_kernel_size,
+                    context_size=config.context_size,  # attention context size
+                    attn_dropout=config.dropout,
+                    ff_dropout=config.dropout,
+                    conv_dropout=config.dropout,
+                )
+            )
+            self.rnn_tr = nn.Sequential(*self.rnn_trL)
+
+        self.out = nn.Linear(config.hidden_dim, config.output_dim, bias=True)
+        self.out_mid = nn.Linear(config.output_dim, config.hidden_dim, bias=True)
+        self.context_size = config.context_size
+        self.input_dim = config.input_dim
+        self.num_layers = config.num_layers
+        self.hidden_dim = config.hidden_dim
+        self.output_dim = config.output_dim
+
+    def forward(self, x: torch.Tensor):
+        x = self.rnn_trL[0](x)
+        for l in range(1, self.num_layers + 1):
+            x = self.rnn_trL[l](x, self.context_size)
+            if l == self.num_layers // 2:
+                x_mid = x.clone()
+                x_mid = self.out(x_mid)
+                x += self.out_mid(nn.Softmax(dim=-1)(x_mid))
+        return x
+
+
+# NOTE: Conformer code is adapated from the following
+# https://github.com/lucidrains/conformer.git
+# helper functions
+
+def calc_same_padding(kernel_size):
+    pad = kernel_size // 2
+    return (pad, pad - (kernel_size + 1) % 2)
+
+
+class Permute(nn.Module):
+    def __init__(self, dims):
+        super().__init__()
+        self.dims = dims
+        
+    def forward(self, x):
+        x = x.permute(self.dims)
+        return x
+
+
+# helper classes
+
+class DepthWiseConv1d(nn.Module):
+    def __init__(self, chan_in, chan_out, kernel_size, padding):
+        super().__init__()
+        self.padding = padding
+        self.conv = nn.Conv1d(chan_in, chan_out, kernel_size, groups = chan_in, bias=False)
+
+    def forward(self, x):
+        x = F.pad(x, self.padding)
+        return self.conv(x)
+
+# attention, feedforward, and conv module
+
+class Scale(nn.Module):
+    def __init__(self, scale, fn):
+        super().__init__()
+        self.fn = fn
+        self.scale = scale
+
+    def forward(self, x, **kwargs):
+        return self.fn(x, **kwargs) * self.scale
+
+class PreNorm(nn.Module):
+    def __init__(self, dim, fn):
+        super().__init__()
+        self.fn = fn
+        self.norm = nn.LayerNorm(dim)
+
+    def forward(self, x, **kwargs):
+        x = self.norm(x)
+        return self.fn(x, **kwargs)
+
+class PreNormAttn(nn.Module):
+    def __init__(self, dim, fn):
+        super().__init__()
+        self.fn = fn
+        self.norm = nn.LayerNorm(dim)
+
+    def forward(self, x, context_size, **kwargs):
+        x = self.norm(x)
+        return self.fn(x, context_size, **kwargs)
+
+class Attention(nn.Module):
+    def __init__(
+        self,
+        dim,
+        heads = 8,
+        dim_head = 64,
+        dropout = 0.,
+        context_size = 200,
+        max_pos_emb = 512
+    ):
+        super().__init__()
+        inner_dim = dim_head * heads
+        self.heads= heads
+        self.dim_head = dim_head
+        self.scale = dim_head ** -0.5
+        self.to_q = nn.Linear(dim, inner_dim, bias = False)
+        self.to_kv = nn.Linear(dim, inner_dim * 2, bias = False)
+        self.to_out = nn.Linear(inner_dim, dim)
+
+        self.max_pos_emb = max_pos_emb
+        self.rel_pos_emb = nn.Embedding(2 * max_pos_emb + 1, dim_head)
+
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x, context_size):
+        device, h, max_pos_emb = x.device, self.heads, self.max_pos_emb
+        bs, n, d = x.shape
+        assert(context_size > 0 and context_size <= max_pos_emb)
+
+        nb = n // context_size
+        nr = n % context_size
+        if nr > 0:
+            y = torch.zeros(x.shape[0], context_size-nr, x.shape[2], device=device)
+            x = torch.cat((x,y), dim=1)
+            nb += 1
+
+        q, k, v = (self.to_q(x), *self.to_kv(x).chunk(2, dim = -1))
+        q, k, v = map(lambda t: t.reshape(bs, nb, context_size, h, -1).transpose(2, 3), (q, k, v))
+        dots = einsum('b m h i d, b m h j d -> b m h i j', q, k) * self.scale
+
+        # shaw's relative positional embedding
+        seq = torch.arange(context_size, device = device)
+        dist = seq.view(-1, 1) - seq.view(1, -1)
+        dist = torch.clamp(dist,-context_size, context_size) + max_pos_emb
+        rel_pos_emb = self.rel_pos_emb(dist).to(q)
+        pos_attn = einsum('b m h c d, c r d -> b m h c r', q, rel_pos_emb) * self.scale
+        dots = dots + pos_attn
+
+        if nr > 0:
+            mask = torch.ones(context_size, context_size, device=device)
+            mask[:nr,:nr] = 0
+            mask_value = -torch.finfo(dots.dtype).max
+            dots[:,-1,:].masked_fill_(mask.bool(), mask_value)
+
+        attn = dots.softmax(dim = -1)
+
+        out = einsum('b m h i j, b m h j d -> b m h i d', attn, v)
+        out = out.transpose(2, 3).reshape(bs, x.shape[1], -1)
+        out = self.to_out(out[:,:n,:])
+        return self.dropout(out)
+
+class FeedForward(nn.Module):
+    def __init__(
+        self,
+        dim,
+        mult = 4,
+        dropout = 0.
+    ):
+        super().__init__()
+        self.net = nn.Sequential(
+            nn.Linear(dim, dim * mult),
+            nn.SiLU(),
+            nn.Dropout(dropout),
+            nn.Linear(dim * mult, dim),
+            nn.Dropout(dropout)
+        )
+
+    def forward(self, x):
+        return self.net(x)
+
+class ConformerConvModule(nn.Module):
+    def __init__(
+        self,
+        dim,
+        causal = False,
+        expansion_factor = 2,
+        kernel_size = 31,
+        dropout = 0.):
+        super().__init__()
+
+        inner_dim = dim * expansion_factor
+        padding = calc_same_padding(kernel_size) if not causal else (kernel_size - 1, 0)
+
+        self.net = nn.Sequential(
+            nn.LayerNorm(dim),
+            Permute(dims=(0, 2, 1)),
+            nn.Conv1d(dim, inner_dim * 2, 1),
+            nn.GLU(dim=1),
+            DepthWiseConv1d(inner_dim, inner_dim, kernel_size = kernel_size, padding = padding),
+            nn.BatchNorm1d(inner_dim) if not causal else nn.Identity(),
+            nn.SiLU(),
+            nn.Conv1d(inner_dim, dim, 1),
+            Permute(dims=(0, 2, 1)),
+            nn.Dropout(dropout)
+        )
+
+    def forward(self, x):
+        return self.net(x)
+
+# Conformer Block
+
+class ConformerBlock(nn.Module):
+    def __init__(
+        self,
+        *,
+        dim,
+        dim_head = 64,
+        heads = 8,
+        ff_mult = 2,
+        conv_expansion_factor = 2,
+        conv_kernel_size = 31,
+        context_size = -1,
+        attn_dropout = 0.,
+        ff_dropout = 0.,
+        conv_dropout = 0.
+    ):
+        super().__init__()
+        self.ff1 = FeedForward(dim = dim, mult = ff_mult, dropout = ff_dropout)
+        self.attn = Attention(dim = dim, dim_head = dim_head, heads = heads, dropout = attn_dropout, context_size = context_size)
+        self.conv = ConformerConvModule(dim = dim, causal = False, expansion_factor = conv_expansion_factor, kernel_size = conv_kernel_size, dropout = conv_dropout)
+        self.ff2 = FeedForward(dim = dim, mult = ff_mult, dropout = ff_dropout)
+
+        self.attn = PreNormAttn(dim, self.attn)
+        self.ff1 = Scale(0.5, PreNorm(dim, self.ff1))
+        self.ff2 = Scale(0.5, PreNorm(dim, self.ff2))
+
+        self.post_norm = nn.LayerNorm(dim)
+
+    def forward(self, x, context_size):
+        x = self.ff1(x) + x
+        x = self.attn(x, context_size) + x
+        x = self.conv(x) + x
+        x = self.ff2(x) + x
+        x = self.post_norm(x)
+        return x
+
+
+
 class GraniteSpeechForConditionalGeneration(PreTrainedModel, GenerationMixin):
+    _supports_cache_class = True
     def __init__(self, config: GraniteSpeechConfig):
         super().__init__(config)
 
-        self.llm = GraniteForCausalLM.from_pretrained(config.llm_name)
+        self.language_model = GraniteForCausalLM.from_pretrained(config.llm_name)
+        # TODO - See if we can use lora layers or this can be moved out to a conditional wrapper
         peft_config = LoraConfig(
             task_type=TaskType.CAUSAL_LM,
             inference_mode=True,
@@ -26,26 +331,40 @@ class GraniteSpeechForConditionalGeneration(PreTrainedModel, GenerationMixin):
             lora_alpha=config.lora_alpha,
             target_modules=config.lora_modules,
         )
-        self.llm = get_peft_model(self.llm, peft_config)
+        self.language_model = get_peft_model(self.language_model, peft_config)
 
+        # TODO - move all of this stuff out
         self.encoder = CTCModel(config.encoder_config)
-
         self.projector = EncoderProjectorQFormer(config.projector_config)
-
         encoder_state_dict = torch.load(
             "data/encoder.pt", map_location="cpu", weights_only=True
         )
-        print(self.encoder.load_state_dict(encoder_state_dict, strict=False))
+        self.encoder.load_state_dict(encoder_state_dict, strict=False)
 
         lora_state_dict = torch.load(
             "data/lora_adapter.pt", map_location="cpu", weights_only=True
         )
-        self.llm.load_state_dict(lora_state_dict, strict=False)
+        self.language_model.load_state_dict(lora_state_dict, strict=False)
 
         projector_state_dict = torch.load(
             "data/projector.pt", map_location="cpu", weights_only=True
         )
         self.projector.load_state_dict(projector_state_dict, strict=True)
+        self.post_init()
+
+    def get_input_embeddings(self):
+        return self.language_model.get_input_embeddings()
+
+    def get_audio_features(self, input_features):
+        # TODO - remove timers, keeping them for now to ensure we don't
+        # add a ton of extra latency while we are porting the model.
+        a = time.time()
+        encoder_embeds = self.encoder(input_features)
+        print("Encoder", time.time() - a, "secs")
+
+        a = time.time()
+        projected_embeds = self.projector(encoder_embeds, None)
+        return projected_embeds
 
     def forward(
         self,
@@ -61,67 +380,90 @@ class GraniteSpeechForConditionalGeneration(PreTrainedModel, GenerationMixin):
         output_attentions: Optional[bool] = None,
         output_hidden_states: Optional[bool] = None,
         return_dict: Optional[bool] = None,
-    ): 
-        if inputs_embeds is None:
-            inputs_embeds = self.prepare_inputs_for_generation(
-                input_ids=input_ids,
-                input_features=input_features,
-                attention_mask=attention_mask,
-            )
-        llm_outputs = self.llm(inputs_embeds=inputs_embeds, 
-                               attention_mask=attention_mask,
-                               past_key_values=past_key_values,
-                               position_ids=position_ids,
-                               labels=labels, 
-                               use_cache=use_cache,
-                               output_attentions=output_attentions,
-                               output_hidden_states=output_hidden_states, 
-                               return_dict=return_dict,
-
-                               )
-        return llm_outputs
-
-    def generate(
-        self,
-        input_ids,
-        inputs_embeds=None,
-        input_features=None,
-        attention_mask=None,
-        **kwargs,
+        logits_to_keep: Union[int, torch.Tensor] = 0,
     ):
-        if inputs_embeds is None:
-            inputs_embeds = self.prepare_inputs_for_generation(
-                input_ids=input_ids,
-                input_features=input_features,
-                attention_mask=attention_mask,
+        # Similar to llava; if we have input IDs, we encode them into embeddings.
+        # On the first pass, we should have no input embeddings, and only input features,
+        # since we need to encode the features into the LLM's embedded output.
+        if (input_ids is None) ^ (inputs_embeds is not None):
+            raise ValueError("You must specify exactly one of input_ids or inputs_embeds")
+
+        if input_features is not None and inputs_embeds is not None:
+            raise ValueError(
+                "You cannot specify both pixel_values and inputs_embeds at the same time, and must specify either one"
             )
-        model_outputs = self.llm.generate(
-            inputs_embeds=inputs_embeds, attention_mask=attention_mask, **kwargs
+
+        if inputs_embeds is None:
+            # Get the base embeddings; set all audio tokens to 0 index
+            # to avoid out of vocabulary issues with the LLM embedding.
+            # Audio features will be masked into is_audio_idx indices later.
+            is_audio_idx = input_ids == self.config.audio_token_index
+            llm_input_ids = input_ids.clone()
+            llm_input_ids[is_audio_idx] = 0
+            inputs_embeds = self.get_input_embeddings()(llm_input_ids)
+
+        if input_features is not None:
+            # Get the audio features from the encoder / projector 
+            audio_features = self.get_audio_features(input_features)
+
+            # Merge the audio features into the LLM embeddings
+            inputs_embeds = self.get_merged_audio_embeddings(
+                input_ids=input_ids,
+                audio_features=audio_features,
+            )
+
+        outputs = self.language_model(
+            inputs_embeds=inputs_embeds, 
+            attention_mask=attention_mask,
+            past_key_values=past_key_values,
+            position_ids=position_ids,
+            labels=labels, 
+            use_cache=use_cache,
+            output_attentions=output_attentions,
+            output_hidden_states=output_hidden_states, 
+            return_dict=return_dict,
         )
-        return model_outputs
+
+        logits = outputs[0]
+
+        return GraniteSpeechCausalLMOutputWithPast(
+            loss=None, # TODO 
+            logits=logits,
+            past_key_values=outputs.past_key_values,
+            hidden_states=outputs.hidden_states,
+            attentions=outputs.attentions,
+        )
+
 
     def prepare_inputs_for_generation(
         self,
         input_ids,
+        past_key_values=None,
+        inputs_embeds=None,
         input_features=None,
+        attention_mask=None,
+        cache_position=None,
+        logits_to_keep=None,
         **kwargs,
     ):
-        a = time.time()
-        encoder_embeds = self.encoder(input_features)
-        print("Encoder", time.time() - a, "secs")
+        # Overwritten -- in specific circumstances we don't want to forward audio inputs to the model
 
-        a = time.time()
-        projected_embeds = self.projector(encoder_embeds, None)
-        print("Projector", time.time() - a, "secs")
-
-        a = time.time()
-        # concatenate embeddings and invoke LLM generate
-        # tokenizer.vocab[self_processor.audio_token]
-        combined_embeds = self.get_merged_audio_embeddings(
-            input_ids=input_ids,
-            audio_features=projected_embeds,
+        model_inputs = self.language_model.prepare_inputs_for_generation(
+            input_ids,
+            past_key_values=past_key_values,
+            inputs_embeds=inputs_embeds,
+            attention_mask=attention_mask,
+            cache_position=cache_position,
+            logits_to_keep=logits_to_keep,
+            **kwargs,
         )
-        return combined_embeds
+
+        # If we're in cached decoding stage, pixel values should be None because input ids
+        # do not contain special audio token anymore Otherwise we need input feature values
+        # to be passed to the model
+        if cache_position[0] == 0:
+            model_inputs["input_features"] = input_features
+        return model_inputs
 
     def get_merged_audio_embeddings(self, input_ids, audio_features):
         """
@@ -134,7 +476,7 @@ class GraniteSpeechForConditionalGeneration(PreTrainedModel, GenerationMixin):
         """
         is_audio_index = input_ids == self.config.audio_token_index
         llm_input_ids = torch.where(is_audio_index, 0, input_ids)
-        inputs_embeds = self.llm.get_input_embeddings()(
+        inputs_embeds = self.language_model.get_input_embeddings()(
             llm_input_ids
         )  # [bsz, # features, hidden size]
